[DEL] 0.0.35 main page & repalced & delete deprecated files
