+++ src/api/mainpage.ts
@@ -1,30 +1,40 @@
-import type { NoticeProps } from "@api/mainpage/noticeData";
+import type { NoticeProps } from "@api/mainpage/commonData";
+import type { BannerProps } from "@api/mainpage/bannerData";
 
 import {
-  mainBannerData,
-  introduceImages,
-  movieImages,
-} from "@api/mainpage/imageData";
-import { noticeTypesData, noticeData } from "@api/mainpage/noticeData";
+  hotelInfos,
+  moviePrices,
+  noticeData,
+  noticeTypesData,
+} from "@api/mainpage/commonData";
 
-/* =============== [ swipers ] =============== */
-const getMainBannerData = (): BannerProps[] => {
-  return mainBannerData;
-};
-type IntroduceImageType = "stay" | "tour" | "reservation" | "customerCenter";
+import {
+  mainTopBannerData,
+  arrivalBannerData,
+  movieBannerData,
+  tourBannerData,
+} from "@api/mainpage/bannerData";
 
-const getIntroduceImages = (
-  type: IntroduceImageType,
-  size: number = 1
-): ImageProps[] => {
-  const returnArray = introduceImages[type];
-  return returnArray.slice(
-    0,
-    returnArray.length < size ? returnArray.length : size
-  );
+const getTopBanners = () => {
+  return mainTopBannerData;
 };
-const getMovieData = (): ImageProps[] => {
-  return movieImages;
+const getArrivalBanners = () => {
+  return arrivalBannerData;
+};
+
+const getHotelInfo = () => {
+  return hotelInfos;
+};
+
+const getTourBanners = () => {
+  return tourBannerData;
+};
+
+const getMovieInfo = () => {
+  return moviePrices;
+};
+const getMovieBanners = () => {
+  return movieBannerData;
 };
 const getNoticeData = (): NoticeProps[] => {
   return noticeData;
@@ -33,11 +43,13 @@ const getNoticeTypeString = (type: "gurye" | "goesan"): string => {
   return noticeTypesData[type];
 };
 
-export type { NoticeProps };
 export {
-  getMainBannerData,
-  getIntroduceImages,
-  getMovieData,
+  getTopBanners,
+  getArrivalBanners,
+  getMovieBanners,
+  getTourBanners,
   getNoticeData,
   getNoticeTypeString,
 };
+export { getHotelInfo, getMovieInfo };
+export type { BannerProps };
