--- src/api/mainpage/imageData.ts
@@ -1,118 +1,0 @@
-const mainBannerData: BannerProps[] = [
-  {
-    key: "swiper-geosan-tour",
-    to: "https://www.naturaldreampark.co.kr/g_gongbang_book.html",
-    image: {
-      src: "https://www.naturaldreampark.co.kr/images/main_banner09.jpg",
-      alt: "괴산 공방견학",
-    },
-    mobileImage: {
-      src: "https://www.naturaldreampark.co.kr/mobile/images/main_banner04m.jpg?ver=3",
-      alt: "괴산 공방견학",
-    },
-  },
-  {
-    key: "swiper-goesan-roum",
-    to: "https://www.naturaldreampark.co.kr/g_facilities_information.html",
-    image: {
-      src: "https://www.naturaldreampark.co.kr/images/main_banner10.jpg",
-      alt: "괴산 로움호텔",
-    },
-    mobileImage: {
-      src: "https://www.naturaldreampark.co.kr/mobile/images/main_banner05m.jpg?ver=3",
-      alt: "괴산 로움호텔",
-    },
-  },
-  {
-    key: "swiper-gurye-sprout-tour",
-    to: "https://www.naturaldreampark.co.kr/reservation_new4.html",
-    image: {
-      src: "https://www.naturaldreampark.co.kr/images/main_banner11.jpg",
-      alt: "구례 새싹농사 체험",
-    },
-    mobileImage: {
-      src: "https://www.naturaldreampark.co.kr/mobile/images/main_banner06m.jpg?ver=3",
-      alt: "구례 새싹농사 체험",
-    },
-  },
-  {
-    key: "swiper-gurye-hue-pension",
-    to: "https://www.naturaldreampark.co.kr/reservation_hue_pension.html",
-    image: {
-      src: "https://www.naturaldreampark.co.kr/images/main_banner12.jpg",
-      alt: "구례 휴펜션",
-    },
-    mobileImage: {
-      src: "https://www.naturaldreampark.co.kr/mobile/images/main_banner07m.jpg?ver=3",
-      alt: "구례 휴펜션",
-    },
-  },
-];
-
-const introduceImages = {
-  stay: [
-    {
-      src: "https://www.naturaldreampark.co.kr/images/slide1_01.jpg",
-      alt: "오직 조합원만을 위한 숙박시설",
-    },
-    {
-      src: "https://www.naturaldreampark.co.kr/images/slide1_02.jpg",
-      alt: "오직 조합원만을 위한 숙박시설",
-    },
-  ],
-  tour: [
-    {
-      src: "https://www.naturaldreampark.co.kr/images/slide2_03.jpg",
-      alt: "공방 견학부터 만들기 체험까지",
-    },
-    {
-      src: "https://www.naturaldreampark.co.kr/images/slide2_02.jpg",
-      alt: "공방 견학부터 만들기 체험까지",
-    },
-    {
-      src: "https://www.naturaldreampark.co.kr/images/slide2_01.jpg",
-      alt: "공방 견학부터 만들기 체험까지",
-    },
-  ],
-  reservation: [
-    {
-      src: "https://www.naturaldreampark.co.kr/images/new_reserve_banner3.jpg",
-      alt: "예약 안내",
-    },
-  ],
-  customerCenter: [
-    {
-      src: "https://www.naturaldreampark.co.kr/images/new_information_banner2.jpg",
-      alt: "자연드림파크 이용약관 및 개인정보 처리방침 변경 안내",
-    },
-  ],
-};
-
-const movieImages: ImageProps[] = [
-  {
-    src: "https://icoop.or.kr/coopmall/ext/nfs/icoop_admin/file/notice/화면 캡처 2025-02-11 183354.jpg",
-    alt: "히트맨2 (15세 관람가 / 118분)",
-  },
-  {
-    src: "https://icoop.or.kr/coopmall/ext/nfs/icoop_admin/file/notice/코난.jpg",
-    alt: "명참정 코난: 14번째 표적 (12세 관람가 / 100분)",
-  },
-  {
-    src: "https://icoop.or.kr/coopmall/ext/nfs/icoop_admin/file/notice/말할수없는비밀.jpg",
-    alt: "말할 수 없는 비밀 (전체관람가 / 103분)",
-  },
-  {
-    src: "https://icoop.or.kr/coopmall/ext/nfs/icoop_admin/file/notice/1화면 캡처 2025-01-30 140911.jpg",
-    alt: "브로큰(15세 관람가 / 99분)",
-  },
-  {
-    src: "https://icoop.or.kr/coopmall/ext/nfs/icoop_admin/file/notice/캡틴아메리카.jpg",
-    alt: "캡틴 아메리카: 브레이브 뉴 월드 (12세 관람가 / 118분)",
-  },
-  {
-    src: "https://icoop.or.kr/coopmall/ext/nfs/icoop_admin/file/notice/%ED%99%94%EB%A9%B4%20%EC%BA%A1%EC%B2%98%202025-02-08%20125505.jpg",
-    alt: "9월 5일: 위험한 특종 (15세 관람가 / 95분)",
-  },
-];
-
-export { mainBannerData, introduceImages, movieImages };
