+++ src/components/MainPage/TourBanner.tsx
@@ -5,7 +5,7 @@ import Container from "@components/Container";
 import MainPageHeading from "@components/MainPage/MainPageHeading";
 import MainPageBanner from "@components/MainPage/MainPageBanner";
 
-import { getTourBanners } from "@api/new-mainpage";
+import { getTourBanners } from "@api/mainpage";
 
 const TourBanner: React.FC = () => {
   return (
